refactory code and update design app
