+++ src-tauri/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "simplecont-app"
-version = "0.0.0"
+version = "1.0.0"
 description = "A Tauri App"
-authors = ["you"]
+authors = ["felipecaninnovaes"]
 license = ""
 repository = ""
 edition = "2021"
