+++ src-tauri/tauri.conf.json
@@ -2,12 +2,12 @@
   "build": {
     "beforeDevCommand": "yarn dev",
     "beforeBuildCommand": "yarn build",
-    "devPath": "http://localhost:1420",
+    "devPath": "http://localhost:1488",
     "distDir": "../dist"
   },
   "package": {
     "productName": "simplecont-app",
-    "version": "0.0.0"
+    "version": "1.0.0"
   },
   "tauri": {
     "allowlist": {
@@ -28,7 +28,7 @@
         "icons/icon.icns",
         "icons/icon.ico"
       ],
-      "identifier": "com.felipecn.sct",
+      "identifier": "com.felipecaninnovaes.sct",
       "longDescription": "",
       "macOS": {
         "entitlements": null,
@@ -55,10 +55,10 @@
     "windows": [
       {
         "fullscreen": false,
-        "height": 600,
+        "height": 690,
         "resizable": false,
-        "title": "simplecont-app",
-        "width": 330
+        "title": "SimpleCount",
+        "width": 666
       }
     ]
   }
