+++ src/styles/pagesBase.css
@@ -1,0 +1,82 @@
+.app {
+    position: absolute;
+    left: 20.99%;
+    right: 20.99%;
+    top: 11.27%;
+    bottom: 11.27%;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    background: #f1f1f1;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 10px;
+}
+
+.itens {
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    margin: 8%;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+.Lines {
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    width: 100%;
+}
+
+input {
+    width: 100%;
+}
+
+input {
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    color: #0f0f0f;
+    background-color: #dbdbdb;
+    transition: border-color 0.25s;
+    box-shadow: 0 2px 2px rgba(0, 0, 0, 0.2);
+}
+
+.cal {
+    
+    background-color: #5C6BB8;
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    color: #ffffff;
+    transition: border-color 0.25s;
+    box-shadow: 0 2px 2px rgba(0, 0, 0, 0.2);
+    width: 100%;
+    align-content: center;
+    align-items: center;
+  }
+  
+.cal {
+    margin-top: 8px;
+    cursor: pointer;
+  }
+  
+.cal:hover {
+    border-color: #396cd8;
+  }
+
+.buttons {
+    align-content: center;
+    align-items: center;
+    width: 100%;
+}
+
+::placeholder {
+    color: #5e5e5e;
+}
